Add teacher staff own profile, user serializable
commit d58bae4cd2b0d2b6ff2a034a5aed2025f1bf88a4
Author: Ashish Kumar <ashishkr23438@gmail.com>
Date:   Sun Nov 1 11:51:10 2020 +0530

    Revert adding guardian in student

commit 2cef2c119f0e04cb1dfaf5197afd57938049edd0
Author: Ashish Kumar <ashishkr23438@gmail.com>
Date:   Sun Nov 1 11:42:23 2020 +0530

    Removed demo folder

commit 1dafbd5c24c565c9ca8c226b4ac97a6c0b3e47fc
Author: Ashish Kumar <ashishkr23438@gmail.com>
Date:   Sun Nov 1 11:40:42 2020 +0530

    Added Teacher & Staff Own Profile

commit 8c96d11a8ce8c9e52de20e3f49912ebfad055165
Author: Ashish Kumar <ashishkr23438@gmail.com>
Date:   Thu Oct 29 01:04:51 2020 +0530

    Made User model Serializable

commit 71040ebd4ccf7fd33b9f1c5f2f94d7ab6b4c66fd
Author: Ashish Kumar <ashishkr23438@gmail.com>
Date:   Thu Oct 29 00:36:57 2020 +0530

    Temp commit

+++ src/main/java/com/dbms/coaching/controllers/UserController.java
@@ -6,6 +6,7 @@ import com.dbms.coaching.dao.UserDao;
 import com.dbms.coaching.dao.UserPhoneNumberDao;
 import com.dbms.coaching.models.User;
 import com.dbms.coaching.models.UserPhoneNumber;
+import com.dbms.coaching.services.SecurityService;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -25,12 +26,14 @@ public class UserController {
     @Autowired
     private UserPhoneNumberDao userPhoneNumberDao;
 
+    @Autowired
+    private SecurityService securityService;
+
     @GetMapping("/admin/users")
     public String usersPortal(Model model) {
         model.addAttribute("title", "Users Portal");
         model.addAttribute("message", "View all the users");
         List<User> users = userDao.getAll();
-        // TODO: Prevent deleting self user
         model.addAttribute("users", users);
         return "user/usersPortal";
     }
@@ -42,14 +45,22 @@ public class UserController {
     }
 
     @GetMapping("/admin/users/{userId}/delete")
-    public ResponseEntity<Integer> deleteUser(@PathVariable("userId") int userId, Model model) {
+    public ResponseEntity<String> deleteUser(@PathVariable("userId") int userId, Model model) {
+        if (userId == securityService.findLoggedInUserId()) {
+            return new ResponseEntity<>("Can't delete own user", HttpStatus.BAD_REQUEST);
+        }
         userDao.delete(userId);
         return new ResponseEntity<>(HttpStatus.OK);
     }
 
-    @PostMapping("/admin/users/{userId}/phoneNumber/add")
+    @PostMapping("/profile/users/{userId}/phoneNumber/add")
     public ResponseEntity<String> addUserPhoneNumber(@PathVariable("userId") int userId,
             @RequestParam String phoneNumber, Model model) {
+        int loggedInUserId = securityService.findLoggedInUserId();
+        String role = securityService.findLoggedInUserRole();
+        if (!role.equals("admin") && loggedInUserId != userId)
+            return new ResponseEntity<>("You are not allowed to perform this action", HttpStatus.BAD_REQUEST);
+
         UserPhoneNumber userPhoneNumber = new UserPhoneNumber(phoneNumber, userId);
         userPhoneNumberDao.save(userPhoneNumber);
         // TODO: Check for Validation
@@ -59,9 +70,14 @@ public class UserController {
         return new ResponseEntity<>(HttpStatus.OK);
     }
 
-    @PostMapping("/admin/users/{userId}/phoneNumber/delete")
-    public ResponseEntity<Integer> deleteUserPhoneNumber(@PathVariable("userId") int userId,
+    @PostMapping("/profile/users/{userId}/phoneNumber/delete")
+    public ResponseEntity<String> deleteUserPhoneNumber(@PathVariable("userId") int userId,
             @RequestParam String phoneNumber, Model model) {
+        int loggedInUserId = securityService.findLoggedInUserId();
+        String role = securityService.findLoggedInUserRole();
+        if (!role.equals("admin") && loggedInUserId != userId)
+            return new ResponseEntity<>("You are not allowed to perform this action", HttpStatus.BAD_REQUEST);
+
         UserPhoneNumber userPhoneNumber = new UserPhoneNumber(phoneNumber, userId);
         userPhoneNumberDao.delete(userPhoneNumber);
         return new ResponseEntity<>(HttpStatus.OK);
